Rakic-Zaposleni Prepravljeno

+++ CirkusApp/CirkusApp1/Entiteti/Zaposleni.cs
@@ -14,7 +14,7 @@ namespace CirkusApp1.Entiteti
         public virtual string Prezime { get; set; }
         public virtual DateTime Datum_Rodj { get; set; }
         public virtual string Mesto_Rodj { get; set; }
-        public virtual int Maticnibr { get; set; }
+        public virtual long Maticnibr { get; set; }
         public virtual Direktor PripadaDirektoru { get; set; }
 
         public Zaposleni()
